requirements.txt
astroid==2.4.2
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
Flask-API==2.0
Flask-Cors==3.0.8
Flask-JWT==0.3.2
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.20.0
pycparser==2.20
PyJWT==1.4.2
pylint==2.5.3
pymongo==3.10.1
python-dotenv==0.13.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
